requirements.txt
fastapi
uvicorn
google-generativeai
openai
tiktoken
python-dotenv
requests
streamlit
python-multipart
